+++ crates/web-sys/src/features/gen_HtmlElement.rs
@@ -629,6 +629,20 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `HtmlElement`*"]
     pub fn set_oninput(this: &HtmlElement, value: Option<&::js_sys::Function>);
+    # [wasm_bindgen (structural , method , getter , js_class = "HTMLElement" , js_name = onbeforeinput)]
+    #[doc = "Getter for the `onbeforeinput` field of this object."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/onbeforeinput)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `HtmlElement`*"]
+    pub fn onbeforeinput(this: &HtmlElement) -> Option<::js_sys::Function>;
+    # [wasm_bindgen (structural , method , setter , js_class = "HTMLElement" , js_name = onbeforeinput)]
+    #[doc = "Setter for the `onbeforeinput` field of this object."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/onbeforeinput)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `HtmlElement`*"]
+    pub fn set_onbeforeinput(this: &HtmlElement, value: Option<&::js_sys::Function>);
     # [wasm_bindgen (structural , method , getter , js_class = "HTMLElement" , js_name = oninvalid)]
     #[doc = "Getter for the `oninvalid` field of this object."]
     #[doc = ""]